refactor: GameRunner 제거
